# pkg/api/rest/docs/swagger.yaml
basePath: /cicada
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.RunScriptReq:
    properties:
      content:
        description: Base64 encoded script content.
        type: string
      mci_id:
        type: string
      ns_id:
        type: string
      vm_id:
        type: string
    type: object
  model.ScriptResult:
    properties:
      error:
        type: string
      is_success:
        type: boolean
      output:
        type: string
    type: object
...
  model.SleepTimeReq:
    properties:
      time:
        description: 'Ex: 300 = 300sec, 1m 30s = 1min 30seconds, 1h 10m 15s = 1hour
          10minutes 15seconds, 1d 1s = 1day 1second, No Input Default: 10s'
        type: string
    type: object
  model.Task:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      id:
        type: string
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
      task_group_id:
        type: string
      workflow_id:
        type: string
    required:
    - id
    - name
    - request_body
    - task_component
    - task_group_id
    - workflow_id
    type: object
...
  model.TaskLog:
    properties:
      content:
        type: string
    type: object
...
info:
  contact: {}
  description: Workflow management module
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Cicada REST API
  version: latest
paths:
...
  /run_script:
    post:
      consumes:
      - application/json
      description: Run script on target with NS ID, MCI ID and VM ID.
      operationId: run-script
      parameters:
      - description: Workflow content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RunScriptReq'
      produces:
      - application/json
      responses:
        "200":
          description: Result of the script running
          schema:
            $ref: '#/definitions/model.ScriptResult'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to run script
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Run script on target
      tags:
      - '[Cicada Task Component]'
  /sleep_time:
    post:
      consumes:
      - application/json
      description: Runs sleep command on cicada and waits for configured time. Wait
        for 10 seconds if time value is not provided.
      operationId: sleep-time
      parameters:
      - description: SleepTime request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SleepTimeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Result of sleep
          schema:
            $ref: '#/definitions/model.ScriptResult'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to run script
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Run sleep command on cicada
      tags:
      - '[Cicada Task Component]'
  /task/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the task directly.
      operationId: get-task-directly
      parameters:
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task Directly
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/task:
    get:
      consumes:
      - application/json
      description: Get a task list of the workflow.
      operationId: list-task
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a task list.
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a task list.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List Task
      tags:
      - '[Workflow]'
  /workflow/{wfId}/task/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the task.
      operationId: get-task
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/task_group/{tgId}/task:
    get:
      consumes:
      - application/json
      description: Get a task list from the task group.
      operationId: list-task-from-task-group
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task group.
        in: path
        name: tgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a task list from the task group.
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a task list from the task group.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List Task from Task Group
      tags:
      - '[Workflow]'
  /workflow/{wfId}/task_group/{tgId}/task/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the task from the task group.
      operationId: get-task-from-task-group
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task group.
        in: path
        name: tgId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task from the task group.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task from the task group.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task from Task Group
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/workflowRun/{wfRunId}/task/{taskId}/taskTryNum/{taskTryNum}/logs:
    get:
      consumes:
      - application/json
      description: Get the task Logs.
      operationId: get-task-logs
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the workflowRunId.
        in: path
        name: wfRunId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      - description: ID of the taskTryNum.
        in: path
        name: taskTryNum
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task Logs.
          schema:
            $ref: '#/definitions/model.TaskLog'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task Logs.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task Logs
      tags:
      - '[Workflow]'
  /workflow/{wfId}/workflowRun/{wfRunId}/task/{taskId}/taskTryNum/{taskTryNum}/logs/download:
    get:
      consumes:
      - application/json
      description: Download the task logs as a file.
      operationId: get-task-logs-download
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the workflowRunId.
        in: path
        name: wfRunId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      - description: ID of the taskTryNum.
        in: path
        name: taskTryNum
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Log file downloaded successfully.
          schema:
            type: file
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task Logs.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Download Task Logs
      tags:
      - '[Workflow]'
...
swagger: "2.0"